Local server shutdown management. Fix some terminal commands.

+++ gns3/servers.py
@@ -148,7 +148,7 @@ class Servers(QtCore.QObject):
         try:
             if sys.platform.startswith("win"):
                 # use the string on Windows
-                self._local_server_proccess = subprocess.Popen(command)
+                self._local_server_proccess = subprocess.Popen(command, creationflags=subprocess.CREATE_NEW_PROCESS_GROUP)
             else:
                 # use arguments on other platforms
                 args = shlex.split(command)
@@ -161,14 +161,17 @@ class Servers(QtCore.QObject):
 
     def stopLocalServer(self, wait=False):
 
-        if self._local_server_proccess:
-            if self._local_server and self._local_server.connected():
-                log.info("sending stop request to the server")
-                self._local_server.send_notification("builtin.stop")
-                self._local_server.close_connection()
+#         if self._local_server and self._local_server.connected():
+#             log.info("sending stop request to the server")
+#             #self._local_server.send_notification("builtin.stop")
+#             self._local_server.close_connection()
+        if self._local_server_proccess and self._local_server_proccess.poll() == None:
+            if sys.platform.startswith("win"):
+                self._local_server_proccess.send_signal(signal.CTRL_BREAK_EVENT)
             else:
                 self._local_server_proccess.send_signal(signal.SIGINT)
-            self._local_server_proccess.wait()
+            if wait:
+                self._local_server_proccess.wait()
 
     def setLocalServer(self, path, host, port):
         """
